Move verify_observer_key into permissions module
Moves verify_observer_key function from records/api.py into permissions/api.py, where it is more relevant as a permissions function.
